Run method in a separate file. Shifted rest related methods to separate file
